Format files

+++ contracts/library/SafeMath.sol
@@ -1,17 +1,19 @@
 pragma solidity 0.5.12;
 
 library SafeMath {
-    function add(uint x, uint y) internal pure returns (uint z) {
+    function add(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x + y) >= x, "ds-math-add-overflow");
     }
-    function sub(uint x, uint y) internal pure returns (uint z) {
+
+    function sub(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x - y) <= x, "ds-math-sub-underflow");
     }
-    function mul(uint x, uint y) internal pure returns (uint z) {
+
+    function mul(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require(y == 0 || (z = x * y) / y == x, "ds-math-mul-overflow");
     }
 
-    function div(uint x, uint y) internal pure returns (uint z) {
+    function div(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require(y > 0, "ds-math-div-overflow");
         z = x / y;
     }
